scraping/requirements.txt
# Python dependencies for the scraping components

scrapy>=2.12.0
supabase>=2.0.0 # Added for Supabase interaction
python-dotenv>=0.19.0 # Added for loading .env files

# Add other dependencies like playwright, etc., as needed
